move filters to their versioned packages
