+++ models/meshmodel/core/policies/rego_policy_relationship.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/layer5io/meshkit/models/meshmodel/registry"
+	"github.com/layer5io/meshkit/models/meshmodel/registry/v1beta1"
 	"github.com/layer5io/meshkit/utils"
 	"github.com/open-policy-agent/opa/rego"
 	"github.com/open-policy-agent/opa/storage"
@@ -25,7 +26,7 @@ func NewRegoInstance(policyDir string, regManager *registry.RegistryManager) (*R
 	var store storage.Store
 
 	ctx := context.Background()
-	registeredRelationships, _, _, err := regManager.GetEntities(&registry.RelationshipFilter{})
+	registeredRelationships, _, _, err := regManager.GetEntities(&v1beta1.RelationshipFilter{})
 	if err != nil {
 		return nil, err
 	}
