+++ models/meshmodel/registry/registry.go
@@ -153,7 +153,7 @@ func (rm *RegistryManager) UpdateEntityStatus(ID string, status string, entityTy
 		return nil
 	}
 }
-
+// to be removed
 func (rm *RegistryManager) GetRegistrants(f *v1beta1.HostFilter) ([]v1beta1.MeshModelHostsWithEntitySummary, int64, error) {
 	var result []v1beta1.MesheryHostSummaryDB
 	var totalcount int64
@@ -215,172 +215,8 @@ func (rm *RegistryManager) GetRegistrants(f *v1beta1.HostFilter) ([]v1beta1.Mesh
 
 func (rm *RegistryManager) GetEntities(f entity.Filter) ([]entity.Entity, int64, int, error) {
 	return f.Get(rm.db)
-
-	// switch filter := f.(type) {
-	// case *v1beta1.ComponentFilter:
-	// 	en := make([]Entity, 0)
-	// 	comps, count, unique := v1beta1.GetMeshModelComponents(rm.db, *filter)
-	// 	for _, comp := range comps {
-	// 		en = append(en, comp)
-	// 	}
-	// 	return en, &count, &unique
-	// case *v1beta1.RelationshipFilter:
-	// 	en := make([]Entity, 0)
-	// 	relationships, count := v1beta1.GetMeshModelRelationship(rm.db, *filter)
-	// 	for _, rel := range relationships {
-	// 		en = append(en, rel)
-	// 	}
-	// 	return en, &count, nil
-	// case *v1beta1.PolicyFilter:
-	// 	en := make([]Entity, 0)
-	// 	policies := v1beta1.GetMeshModelPolicy(rm.db, *filter)
-	// 	for _, pol := range policies {
-	// 		en = append(en, pol)
-	// 	}
-	// 	return en, nil, nil
-	// default:
-	// 	return nil, nil, nil
-	// }
 }
 
-func (rm *RegistryManager) GetModels(db *database.Handler, f entity.Filter) ([]v1beta1.Model, int64, int) {
-	var m []v1beta1.Model
-	type modelWithCategories struct {
-		v1beta1.ModelDB
-		v1beta1.CategoryDB
-		Registry
-		v1beta1.Hostv1beta1
-	}
-
-	countUniqueModels := func(models []modelWithCategories) int {
-		set := make(map[string]struct{})
-		for _, model := range models {
-			key := model.ModelDB.Name + "@" + model.ModelDB.Version
-			if _, ok := set[key]; !ok {
-				set[key] = struct{}{}
-			}
-		}
-		return len(set)
-	}
-
-	var modelWithCategoriess []modelWithCategories
-	finder := db.Model(&v1beta1.ModelDB{}).
-		Select("model_dbs.*, category_dbs.*", "registries.*", "hosts.*").
-		Joins("JOIN category_dbs ON model_dbs.category_id = category_dbs.id").
-		Joins("JOIN registries ON registries.entity = model_dbs.id").
-		Joins("JOIN hosts ON hosts.id = registries.registrant_id")
-
-	// total count before pagination
-	var count int64
-
-	// include components and relationships in response body
-	var includeComponents, includeRelationships bool
-
-	if mf, ok := f.(*ModelFilter); ok {
-		if mf.Greedy {
-			if mf.Name != "" && mf.DisplayName != "" {
-				finder = finder.Where("model_dbs.name LIKE ? OR model_dbs.display_name LIKE ?", "%"+mf.Name+"%", "%"+mf.DisplayName+"%")
-			} else if mf.Name != "" {
-				finder = finder.Where("model_dbs.name LIKE ?", "%"+mf.Name+"%")
-			} else if mf.DisplayName != "" {
-				finder = finder.Where("model_dbs.display_name LIKE ?", "%"+mf.DisplayName+"%")
-			}
-		} else {
-			if mf.Name != "" {
-				finder = finder.Where("model_dbs.name = ?", mf.Name)
-			}
-			if mf.DisplayName != "" {
-				finder = finder.Where("model_dbs.display_name = ?", mf.DisplayName)
-			}
-		}
-		if mf.Annotations == "true" {
-			finder = finder.Where("model_dbs.metadata->>'isAnnotation' = true")
-		} else if mf.Annotations == "false" {
-			finder = finder.Where("model_dbs.metadata->>'isAnnotation' = false")
-		}
-		if mf.Version != "" {
-			finder = finder.Where("model_dbs.version = ?", mf.Version)
-		}
-		if mf.Category != "" {
-			finder = finder.Where("category_dbs.name = ?", mf.Category)
-		}
-		if mf.Registrant != "" {
-			finder = finder.Where("hosts.hostname = ?", mf.Registrant)
-		}
-		if mf.Annotations == "true" {
-			finder = finder.Where("model_dbs.metadata->>'isAnnotation' = true")
-		} else if mf.Annotations == "false" {
-			finder = finder.Where("model_dbs.metadata->>'isAnnotation' = false")
-		}
-		if mf.OrderOn != "" {
-			if mf.Sort == "desc" {
-				finder = finder.Order(clause.OrderByColumn{Column: clause.Column{Name: mf.OrderOn}, Desc: true})
-			} else {
-				finder = finder.Order(mf.OrderOn)
-			}
-		} else {
-			finder = finder.Order("display_name")
-		}
-
-		finder.Count(&count)
-
-		if mf.Limit != 0 {
-			finder = finder.Limit(mf.Limit)
-		}
-		if mf.Offset != 0 {
-			finder = finder.Offset(mf.Offset)
-		}
-		if mf.Status != "" {
-			finder = finder.Where("model_dbs.status = ?", mf.Status)
-		}
-		includeComponents = mf.Components
-		includeRelationships = mf.Relationships
-	}
-	err := finder.
-		Scan(&modelWithCategoriess).Error
-	if err != nil {
-		fmt.Println(modelWithCategoriess)
-		fmt.Println(err.Error()) //for debugging
-	}
-
-	for _, modelDB := range modelWithCategoriess {
-		reg := v1beta1.Hostv1beta1{}
-		model := modelDB.ModelDB.GetModel(modelDB.CategoryDB.GetCategory(db), reg)
-		// host := rm.GetRegistrant(model)
-		// model.HostID = host.ID
-		// model.HostName = host.Hostname
-		// model.DisplayHostName = host.Hostname
-
-		if includeComponents {
-			var components []v1beta1.ComponentDefinitionDB
-			finder := db.Model(&v1beta1.ComponentDefinitionDB{}).
-				Select("component_definition_dbs.id, component_definition_dbs.kind,component_definition_dbs.display_name, component_definition_dbs.api_version, component_definition_dbs.metadata").
-				Where("component_definition_dbs.model_id = ?", model.ID)
-			if err := finder.Scan(&components).Error; err != nil {
-				fmt.Println(err)
-			}
-			// model.Components = components
-		}
-		if includeRelationships {
-			var relationships []v1alpha2.RelationshipDefinitionDB
-			finder := db.Model(&v1alpha2.RelationshipDefinitionDB{}).
-				Select("relationship_definition_dbs.*").
-				Where("relationship_definition_dbs.model_id = ?", model.ID)
-			if err := finder.Scan(&relationships).Error; err != nil {
-				fmt.Println(err)
-			}
-			// model.Relationships = relationships
-		}
-
-		m = append(m, model)
-	}
-	return m, count, countUniqueModels(modelWithCategoriess)
-}
-
-func (rm *RegistryManager) GetCategories(db *database.Handler, f entity.Filter) ([]entity.Entity, int64) {
-	cat, count, _, _ := f.Get(rm.db)
-	return cat, count
-}
 func (rm *RegistryManager) GetRegistrant(e entity.Entity) v1beta1.Hostv1beta1 {
 	eID := e.GetID()
 	var reg Registry
