+++ models/meshmodel/registry/v1beta1/component_filter.go
@@ -1,4 +1,4 @@
-package registry
+package v1beta1
 
 import (
 	"fmt"
